+++ app/app-controllers/welcome-screen/welcomeModalController.js
@@ -99,21 +99,23 @@ function(
                 // Welcome modal should be triggered when installationID changes
                 // NOTE: appInfo.installationID appears to be different on each app run
                 // on the iOS Simulator, but on real devices they will persist.
-                if (appInfo.installationID !== previousInstallationID || params.forced) {
-                    return new Promise(function(resolve, reject) {
-                        self.isShowing = true;
-                        self.modalView.show({animated: true});
+                return new Promise(function(resolve, reject) {
+                    if (appInfo.installationID !== previousInstallationID || params.forced) {
+                            self.isShowing = true;
+                            self.modalView.show({animated: true});
 
-                        // Promise will be resolved when welcome modal is dismissed
-                        AppEvents.on(AppEvents.names.welcomeHidden, function() {
-                            self._secureStore.set('installationID', appInfo.installationID);
-                            resolve();
-                        });
-                        AppEvents.trigger(AppEvents.names.welcomeShown);
-                    });
-                }
+                            // Promise will be resolved when welcome modal is dismissed
+                            AppEvents.on(AppEvents.names.welcomeHidden, function() {
+                                self._secureStore.set('installationID', appInfo.installationID);
+                                resolve();
+                            });
+                            AppEvents.trigger(AppEvents.names.welcomeShown);
+                    } else {
+                        reject();
+                    }
+                });
             }
-        )
+        );
     };
 
     WelcomeModalController.prototype.hide = function() {