+++ app/app.js
@@ -115,11 +115,11 @@ window.run = function() {
         };
 
         var initMainLayout = function() {
-            return createLayout().then(function(menuController) {
+            return createLayout().then(function(layoutController) {
                 // Deep linking services will enable deep linking on startup
                 // and while running it will open the deep link in the current
                 // active tab
-                deepLinkingServices = new DeepLinkingServices(menuController);
+                deepLinkingServices = new DeepLinkingServices(layoutController);
             });
         };