+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## To be released
 - Welcome modal now relies on installationId to determine whether to show
+- Add Requirements to README, and "engines" field to package.json
 
 ## v0.15.0
 - [iOS] Remove update bundle version build phase