+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## To be released
+- Welcome modal now relies on installationId to determine whether to show
 - Add Requirements to README, and "engines" field to package.json
 
 ## v0.15.0